requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-core
google-genai>=0.3.0
python-dotenv
psycopg2-binary
pydantic
pandas
jupyter 
ipykernel
streamlit
numpy
fastapi
uvicorn